refactor: move util.ts into util folder. separate clipboard/messages to own files. don't export default
