+++ src/actions/commands/actions.ts
@@ -3,7 +3,8 @@ import * as vscode from 'vscode';
 import { RecordedState } from '../../state/recordedState';
 import { ReplaceState } from '../../state/replaceState';
 import { VimState } from '../../state/vimState';
-import { allowVSCodeToPropagateCursorUpdatesAndReturnThem, Clipboard } from '../../util';
+import { allowVSCodeToPropagateCursorUpdatesAndReturnThem } from '../../util/util';
+import { Clipboard } from '../../util/clipboard';
 import { FileCommand } from './../../cmd_line/commands/file';
 import { OnlyCommand } from './../../cmd_line/commands/only';
 import { QuitCommand } from './../../cmd_line/commands/quit';
