+++ src/util/clipboard.ts
@@ -1,0 +1,16 @@
+import { logger } from './logger';
+import * as clipboardy from 'clipboardy';
+
+export class Clipboard {
+  public static Copy(text: string) {
+    try {
+      clipboardy.writeSync(text);
+    } catch (e) {
+      logger.error(e, `Clipboard: Error copying to clipboard. Error=${e}`);
+    }
+  }
+
+  public static Paste(): string {
+    return clipboardy.readSync();
+  }
+}
