+++ src/util/logger.ts
@@ -2,10 +2,8 @@ import { configuration } from '../configuration/configuration';
 import * as winston from 'winston';
 import { ConsoleForElectron } from 'winston-console-for-electron';
 
-const logger = winston.createLogger({
+export const logger = winston.createLogger({
   level: configuration.debug.loggingLevel,
   format: winston.format.simple(),
   transports: [new ConsoleForElectron()],
 });
-
-export default logger;
