+++ test/multicursor.test.ts
@@ -3,8 +3,6 @@ import * as assert from 'assert';
 import * as vscode from 'vscode';
 import { getAndUpdateModeHandler } from '../extension';
 import { ModeHandler } from '../src/mode/modeHandler';
-import { TextEditor } from '../src/textEditor';
-import * as util from '../src/util';
 import { getTestingFunctions } from './testSimplifier';
 import {
   assertEqual,
